refactor: 당첨번호, 당첨결과 도메인 분리 fix: 로또 번호 중복으로 나오는 오류 수정
